clear useless code

+++ src/api/remove.php
@@ -1,23 +1,26 @@
 <?php
+
+// global settings, database connection and functions
 include_once '../config/pdo.php';
 include_once '../core/uploadSettings.php';
-
 include_once 'filesRepo.php';
 
 $id = intval($_GET['id']);
 
 $file = findFile($pdo, $id);
 
-if(!$file) {
+if (!$file) {
   header('Location: /');
 }
 
-if(is_file(UPLOAD_DIR . $file['filename'])) {
+if (is_file(UPLOAD_DIR . $file['filename'])) {
   unlink(UPLOAD_DIR . $file['filename']);
 }
 
-$removed = deleteFile($pdo, $id);
-if ($removed) {
-  header('Location: /');
-  exit();
+if (!is_file(UPLOAD_DIR . $file['filename'])) {
+  $removed = deleteFile($pdo, $id);
+  if ($removed) {
+    header('Location: /');
+    exit();
+  }
 }
